+++ variables.tf
@@ -53,27 +53,26 @@ variable "subnet_ids" {
 
 // http://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-optimized_AMI_launch_latest.html
 variable "ecs_optimized_amis" {
-  description = "List with ecs optimized images per region, last updated on: 2018-06-30 (2018.03.a)."
+  description = "List with ecs optimized images per region, last updated on: 2018-08-31 (2018.03.e)."
   type        = "map"
 
   default = {
-    us-east-1      = "ami-5253c32d" # US East (N. Virginia)
-    us-east-2      = "ami-956e52f0" # US East (Ohio)
-    us-west-1      = "ami-6b81980b" # US West (N. California)
-    us-west-2      = "ami-d2f489aa" # US West (Oregon)
-    ca-central-1   = "ami-da6cecbe" # Canada (Central)
-    eu-west-1      = "ami-c91624b0" # EU (Ireland)
-    eu-west-2      = "ami-3622cf51" # EU (London)
-    eu-west-3      = "ami-ca75c4b7" # EU (Paris)
-    eu-central-1   = "ami-10e6c8fb" # EU (Frankfurt)
-    ap-northeast-1 = "ami-f3f8098c" # Asia Pacific (Tokyo)
-    ap-northeast-2 = "ami-7c69c112" # Asia Pacific (Seoul)
-    ap-southeast-1 = "ami-b75a6acb" # Asia Pacific (Singapore)
-    ap-southeast-2 = "ami-bc04d5de" # Asia Pacific (Sydney)
-    ca-central-1   = "ami-da6cecbe" # Canada (Central)
-    ap-south-1     = "ami-c7072aa8" # Asia Pacific (Mumbai)
-    sa-east-1      = "ami-a1e2becd" # South America (São Paulo)
-    us-gov-west-1  = "ami-03920462" # GovCloud
+    us-east-1      = "ami-00129b193dc81bc31" # US East (N. Virginia)
+    us-east-2      = "ami-028a9de0a7e353ed9" # US East (Ohio)
+    us-west-1      = "ami-0d438d09af26c9583" # US West (N. California)
+    us-west-2      = "ami-00d4f478"          # US West (Oregon)
+    ca-central-1   = "ami-192fa27d"          # Canada (Central)
+    eu-west-1      = "ami-0af844a965e5738db" # EU (Ireland)
+    eu-west-2      = "ami-a44db8c3"          # EU (London)
+    eu-west-3      = "ami-07da674f0655ef4e1" # EU (Paris)
+    eu-central-1   = "ami-0291ba887ba0d515f" # EU (Frankfurt)
+    ap-northeast-1 = "ami-0041c416aa23033a2" # Asia Pacific (Tokyo)
+    ap-northeast-2 = "ami-047d2a61f94f862dc" # Asia Pacific (Seoul)
+    ap-southeast-1 = "ami-091bf462afdb02c60" # Asia Pacific (Singapore)
+    ap-southeast-2 = "ami-0092e55c70015d8c3" # Asia Pacific (Sydney)
+    ap-south-1     = "ami-0c179ca015d301829" # Asia Pacific (Mumbai)
+    sa-east-1      = "ami-0018ff8ee48970ac3" # South America (São Paulo)
+    us-gov-west-1  = "ami-c6079ba7"          # GovCloud
   }
 }